pyproject.toml
[project]
name = "ucam-transfer-learning"
version = "0.1.0"
description = ""
authors = [
    {name = "Sangwon Lee", email = "lsw91.main@gmail.com"}
]

requires-python = ">=3.10,<3.11"
dependencies = [
    "torch==2.0.0",
    "torchvision==0.15.1",
    "torchaudio==2.0.1",
    "numpy==1.26.4",
    "statsmodels==0.14.1",
    "pandas==2.2.1",
    "scikit-learn==1.4.1.post1",
    "bottleneck==1.3.8",
    "notebook==7.1.2",
    "einops==0.7.0",
    "mlflow==2.11.3",
    "u8darts[torch]==0.28.0",
    "jinja2==3.1.3",
    "psutil==5.9.8",
    "openpyxl==3.1.2",
    "hydra-core==1.3.2",
    "hydra-joblib-launcher==1.2.0",
    "optuna==3.6.1",
    "optuna-dashboard==0.15.1",
    "optuna-integration==3.6",
    "ipywidgets>=8.1.5",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=3.3.1",
    "pytest>=7.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
index-strategy = "unsafe-best-match"
link-mode = "symlink"
native-tls = true
python-preference = "only-managed"

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu117"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch" }
torchvision = { index = "pytorch" }
torchaudio = { index = "pytorch" }
